Add files via upload
This is a markdown file that I started.   As I see it, to include the latitude and longitude coordinates to the schools in the AcctCurricula database, we would have to join the databases and then loop through each row looking for a matching school name, then capture latitude and longitude coordinates.